move cpp to util.cpp
